# app/python-flask-server-generated/python-flask-server/swagger_server/models/assign_as_mentor.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class AssignAsMentor(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, mentor: int=None, project: int=None):  # noqa: E501
        """AssignAsMentor - a model defined in Swagger

        :param mentor: The mentor of this AssignAsMentor.  # noqa: E501
        :type mentor: int
        :param project: The project of this AssignAsMentor.  # noqa: E501
        :type project: int
        """
        self.swagger_types = {
            'mentor': int,
            'project': int
        }

        self.attribute_map = {
            'mentor': 'mentor',
            'project': 'project'
        }

        self._mentor = mentor
        self._project = project

    @classmethod
    def from_dict(cls, dikt) -> 'AssignAsMentor':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The assign_as_mentor of this AssignAsMentor.  # noqa: E501
        :rtype: AssignAsMentor
        """
        return util.deserialize_model(dikt, cls)

    @property
    def mentor(self) -> int:
        """Gets the mentor of this AssignAsMentor.


        :return: The mentor of this AssignAsMentor.
        :rtype: int
        """
        return self._mentor

    @mentor.setter
    def mentor(self, mentor: int):
        """Sets the mentor of this AssignAsMentor.


        :param mentor: The mentor of this AssignAsMentor.
        :type mentor: int
        """

        self._mentor = mentor

    @property
    def project(self) -> int:
        """Gets the project of this AssignAsMentor.


        :return: The project of this AssignAsMentor.
        :rtype: int
        """
        return self._project

    @project.setter
    def project(self, project: int):
        """Sets the project of this AssignAsMentor.


        :param project: The project of this AssignAsMentor.
        :type project: int
        """

        self._project = project