# app/python-flask-server-generated/python-flask-server/swagger_server/models/assign_users.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class AssignUsers(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, project: int=None, users: List[int]=None):  # noqa: E501
        """AssignUsers - a model defined in Swagger

        :param project: The project of this AssignUsers.  # noqa: E501
        :type project: int
        :param users: The users of this AssignUsers.  # noqa: E501
        :type users: List[int]
        """
        self.swagger_types = {
            'project': int,
            'users': List[int]
        }

        self.attribute_map = {
            'project': 'project',
            'users': 'users'
        }

        self._project = project
        self._users = users

    @classmethod
    def from_dict(cls, dikt) -> 'AssignUsers':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The assign_users of this AssignUsers.  # noqa: E501
        :rtype: AssignUsers
        """
        return util.deserialize_model(dikt, cls)

    @property
    def project(self) -> int:
        """Gets the project of this AssignUsers.


        :return: The project of this AssignUsers.
        :rtype: int
        """
        return self._project

    @project.setter
    def project(self, project: int):
        """Sets the project of this AssignUsers.


        :param project: The project of this AssignUsers.
        :type project: int
        """

        self._project = project

    @property
    def users(self) -> List[int]:
        """Gets the users of this AssignUsers.


        :return: The users of this AssignUsers.
        :rtype: List[int]
        """
        return self._users

    @users.setter
    def users(self, users: List[int]):
        """Sets the users of this AssignUsers.


        :param users: The users of this AssignUsers.
        :type users: List[int]
        """

        self._users = users