# app/python-flask-server-generated/python-flask-server/swagger_server/swagger/swagger.yaml
---
swagger: "2.0"
info:
  description: "Project Management API"
  version: "1.0.0"
  title: "Polls"
host: "djangoapp.public.cy.devfactory.com"
basePath: "/polls"
schemes:
- "http"
paths:
  /:
    get:
      summary: "Root Page"
      operationId: "root_get"
      parameters: []
      responses:
        200:
          description: "This is the expected Response"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
  /user:
    post:
      summary: "Create New User"
      operationId: "user_post"
      parameters:
      - in: "body"
        name: "name"
        description: "Name of the User"
        required: false
        schema:
          $ref: "#/definitions/user"
      responses:
        201:
          description: "Success"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
  /project:
    post:
      summary: "Create New Project"
      operationId: "project_post"
      parameters:
      - in: "body"
        name: "name"
        description: "Name of the Project"
        required: false
        schema:
          $ref: "#/definitions/project"
      responses:
        201:
          description: "Success"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
  /project/assign:
    post:
      summary: "Assign a Project to multiple users"
      operationId: "project_assign_post"
      parameters:
      - in: "body"
        name: "project_id"
        description: "ID of Project and list of users"
        required: false
        schema:
          $ref: "#/definitions/assign"
      responses:
        201:
          description: "Success"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
  /user/assign-as-mentor:
    post:
      summary: "Assign Mentor to a Project"
      operationId: "user_assign_as_mentor_post"
      parameters:
      - in: "body"
        name: "mentor"
        description: "ID of Mentor and Project"
        required: false
        schema:
          $ref: "#/definitions/assign_as_mentor"
      responses:
        201:
          description: "Success"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
  /project/assign-users:
    post:
      summary: "Assign users to a project"
      operationId: "project_assign_users_post"
      parameters:
      - in: "body"
        name: "users"
        description: "ID of Project and List of IDs of Users"
        required: false
        schema:
          $ref: "#/definitions/assign_users"
      responses:
        201:
          description: "Success"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
  /user/{user}/mentees:
    get:
      summary: "Get Mentees of a User"
      operationId: "user_user_mentees_get"
      parameters:
      - name: "user"
        in: "path"
        description: "ID of User"
        required: true
        type: "integer"
      responses:
        201:
          description: "Success"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
  /user/{user}/projects-mentoring:
    get:
      summary: "Get list of projects mentored by a user"
      operationId: "user_user_projects_mentoring_get"
      parameters:
      - name: "user"
        in: "path"
        description: "ID of User"
        required: true
        type: "integer"
      responses:
        201:
          description: "Success"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
  /project/{project}/get-users-mentors:
    get:
      summary: "Get the users and mentors in a project"
      operationId: "project_project_get_users_mentors_get"
      parameters:
      - name: "project"
        in: "path"
        description: "ID of Project"
        required: true
        type: "integer"
      responses:
        201:
          description: "Success"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
  /project/{project}/get-mentors:
    get:
      summary: "Get the mentors of a project"
      operationId: "project_project_get_mentors_get"
      parameters:
      - name: "project"
        in: "path"
        description: "ID of Project"
        required: true
        type: "integer"
      responses:
        201:
          description: "Success"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
definitions:
  user:
    type: "object"
    properties:
      name:
        type: "string"
  project:
    type: "object"
    properties:
      name:
        type: "string"
  assign:
    type: "object"
    properties:
      project_id:
        type: "integer"
      users:
        type: "array"
        items:
          type: "integer"
  assign_as_mentor:
    type: "object"
    properties:
      mentor:
        type: "integer"
      project:
        type: "integer"
  assign_users:
    type: "object"
    properties:
      project:
        type: "integer"
      users:
        type: "array"
        items:
          type: "integer"